review feedback:
- rename a couple files in lambda-common to be more descriptive
- add comments to duplicated jwt.ts files in lambdas describing purpose
