+++ packages/lambda-common/src/constants.ts
@@ -1,0 +1,2 @@
+export const pocketImageCache =
+  'https://pocket-image-cache.com/x/filters:format(jpeg):quality(100):no_upscale():strip_exif()/';
