+++ packages/lambda-common/src/index.ts
@@ -1,3 +1,4 @@
 export * from './jwt';
 export * from './testHelpers';
+export * from './types';
 export * from './validators';
