+++ packages/lambda-common/src/jwt.ts
@@ -9,29 +9,8 @@ const jwt = require('jsonwebtoken');
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const jwkToPem = require('jwk-to-pem');
 
+import { JwtConfig, JwtPayload } from './types';
 import config from './config';
-
-// this is the required subset of properties required by the admin-api gateway
-// https://github.com/Pocket/admin-api/blob/a0bb468cece3ba5bc1a00e1098652a49d433a81d/src/jwtUtils.ts#L98
-// https://github.com/Pocket/curation-tools-data-sync/blob/main/curation-authors-backfill/jwt.ts
-type JwtPayload = {
-  'custom:groups': string; // json enconded string - array of groups
-  aud: string;
-  exp: number;
-  iat: number; //timestamp
-  identities: { userId: string }[];
-  iss: string;
-  name: string;
-};
-
-// config necessary to build payload for signing a JWT
-export type JwtConfig = {
-  aud: string;
-  groups: string[];
-  iss: string;
-  name: string;
-  userId: string;
-};
 
 // reference to secrets manager client
 let smClient: SecretsManagerClient;
